Move client/runtime conn specific code (contributes to #776)
