+++ client/src/fabric/FabricConnection.ts
@@ -14,11 +14,8 @@
 'use strict';
 
 import * as Client from 'fabric-client';
-import * as ClientCA from 'fabric-ca-client';
-import { Gateway, Network, Contract, GatewayOptions, FileSystemWallet, IdentityInfo } from 'fabric-network';
-import { PackageRegistryEntry } from '../packages/PackageRegistryEntry';
-import * as fs from 'fs-extra';
-import { LogType, OutputAdapter } from '../logging/OutputAdapter';
+import { Gateway, GatewayOptions, FileSystemWallet, IdentityInfo } from 'fabric-network';
+import { OutputAdapter } from '../logging/OutputAdapter';
 import { ConsoleOutputAdapter } from '../logging/ConsoleOutputAdapter';
 import { FabricWallet } from './FabricWallet';
 import { URL } from 'url';
@@ -28,11 +25,10 @@ export abstract class FabricConnection {
 
     public identityName: string;
     public wallet: FabricWalletRegistryEntry;
+    protected outputAdapter: OutputAdapter;
+    protected gateway: Gateway = new Gateway();
 
     private mspid: string;
-    private gateway: Gateway = new Gateway();
-    private networkIdProperty: boolean;
-    private outputAdapter: OutputAdapter;
     private discoveryAsLocalhost: boolean;
     private discoveryEnabled: boolean;
 
@@ -43,10 +39,6 @@ export abstract class FabricConnection {
         } else {
             this.outputAdapter = outputAdapter;
         }
-    }
-
-    public isIBPConnection(): boolean {
-        return this.networkIdProperty;
     }
 
     public abstract async connect(wallet: FabricWallet, identityName: string): Promise<void>;
@@ -64,30 +56,6 @@ export abstract class FabricConnection {
         return peerNames;
     }
 
-    public getPeer(name: string): Client.Peer {
-        console.log('getPeer', name);
-        const allPeers: Array<Client.Peer> = this.getAllPeers();
-
-        return allPeers.find((peer: Client.Peer) => {
-            return peer.getName() === name;
-        });
-    }
-
-    public async getOrganizations(channelName: string): Promise<any[]> {
-        console.log('getOrganizations', channelName);
-        const network: Network = await this.gateway.getNetwork(channelName);
-        const channel: Client.Channel = network.getChannel();
-        const orgs: any[] = channel.getOrganizations();
-        return orgs;
-    }
-
-    public getAllCertificateAuthorityNames(): Array<string> {
-        const client: Client = this.gateway.getClient();
-        const certificateAuthority: any = client.getCertificateAuthority();
-        const certificateAuthorityName: string = certificateAuthority.getCaName();
-        return [certificateAuthorityName];
-    }
-
     public async getAllChannelsForPeer(peerName: string): Promise<Array<string>> {
         console.log('getAllChannelsForPeer', peerName);
         // TODO: update this when not just using admin
@@ -103,32 +71,6 @@ export abstract class FabricConnection {
         return channelNames.sort();
     }
 
-    public async getInstalledChaincode(peerName: string): Promise<Map<string, Array<string>>> {
-        console.log('getInstalledChaincode', peerName);
-        const installedChainCodes: Map<string, Array<string>> = new Map<string, Array<string>>();
-        const peer: Client.Peer = this.getPeer(peerName);
-        let chaincodeResponse: Client.ChaincodeQueryResponse;
-        try {
-            chaincodeResponse = await this.gateway.getClient().queryInstalledChaincodes(peer);
-        } catch (error) {
-            if (error.message && error.message.match(/access denied/)) {
-                // Not allowed to do this as we're probably not an administrator.
-                // This is probably not the end of the world, so return the empty map.
-                return installedChainCodes;
-            }
-            throw error;
-        }
-        chaincodeResponse.chaincodes.forEach((chaincode: Client.ChaincodeInfo) => {
-            if (installedChainCodes.has(chaincode.name)) {
-                installedChainCodes.get(chaincode.name).push(chaincode.version);
-            } else {
-                installedChainCodes.set(chaincode.name, [chaincode.version]);
-            }
-        });
-
-        return installedChainCodes;
-    }
-
     public async getInstantiatedChaincode(channelName: string): Promise<Array<{ name: string, version: string }>> {
         console.log('getInstantiatedChaincode');
         const instantiatedChaincodes: Array<any> = [];
@@ -141,256 +83,8 @@ export abstract class FabricConnection {
         return instantiatedChaincodes;
     }
 
-    public async installChaincode(packageRegistryEntry: PackageRegistryEntry, peerName: string): Promise<void> {
-        const peer: Client.Peer = this.getPeer(peerName);
-        const pkgBuffer: Buffer = await fs.readFile(packageRegistryEntry.path);
-        const installRequest: Client.ChaincodePackageInstallRequest = {
-            targets: [peer],
-            chaincodePackage: pkgBuffer,
-            txId: this.gateway.getClient().newTransactionID()
-        };
-        const response: Client.ProposalResponseObject = await this.gateway.getClient().installChaincode(installRequest);
-        const proposalResponse: Client.ProposalResponse | Error = response[0][0];
-        if (proposalResponse instanceof Error) {
-            throw proposalResponse;
-        } else if (proposalResponse.response.status !== 200) {
-            throw new Error(proposalResponse.response.message);
-        }
-    }
-
-    public async instantiateChaincode(name: string, version: string, channelName: string, fcn: string, args: Array<string>): Promise<any> {
-
-        const transactionId: Client.TransactionId = this.gateway.getClient().newTransactionID();
-        const instantiateRequest: Client.ChaincodeInstantiateUpgradeRequest = {
-            chaincodeId: name,
-            chaincodeVersion: version,
-            txId: transactionId,
-            fcn: fcn,
-            args: args
-        };
-
-        const network: Network = await this.gateway.getNetwork(channelName);
-        const channel: Client.Channel = network.getChannel();
-
-        const instantiatedChaincode: Array<any> = await this.getInstantiatedChaincode(channelName);
-
-        const foundChaincode: any = this.getChaincode(name, instantiatedChaincode);
-
-        let proposalResponseObject: Client.ProposalResponseObject;
-
-        let message: string;
-
-        if (foundChaincode) {
-            throw new Error('The name of the contract you tried to instantiate is already instantiated');
-        } else {
-            message = `Instantiating with function: '${fcn}' and arguments: '${args}'`;
-            this.outputAdapter.log(LogType.INFO, undefined, message);
-            proposalResponseObject = await channel.sendInstantiateProposal(instantiateRequest);
-        }
-
-        const contract: Contract = network.getContract(name);
-        const transaction: any = (contract as any).createTransaction('dummy');
-
-        const responses: any = transaction['_validatePeerResponses'](proposalResponseObject[0]);
-
-        const txId: any = transactionId.getTransactionID();
-        const eventHandlerOptions: any = (contract as any).getEventHandlerOptions();
-        const eventHandler: any = transaction['_createTxEventHandler'](txId, network, eventHandlerOptions);
-
-        if (!eventHandler) {
-            throw new Error('Failed to create an event handler');
-        }
-
-        await eventHandler.startListening();
-
-        const transactionRequest: Client.TransactionRequest = {
-            proposalResponses: proposalResponseObject[0] as Client.ProposalResponse[],
-            proposal: proposalResponseObject[1],
-            txId: transactionId
-        };
-
-        // Submit the endorsed transaction to the primary orderers.
-        const response: Client.BroadcastResponse = await network.getChannel().sendTransaction(transactionRequest);
-
-        if (response.status !== 'SUCCESS') {
-            const msg: string = `Failed to send peer responses for transaction ${transactionId.getTransactionID()} to orderer. Response status: ${response.status}`;
-            eventHandler.cancelListening();
-            throw new Error(msg);
-        }
-
-        await eventHandler.waitForEvents();
-        // return the payload from the invoked chaincode
-        let result: any = null;
-        if (responses && responses.validResponses[0].response.payload.length > 0) {
-            result = responses.validResponses[0].response.payload;
-        }
-
-        eventHandler.cancelListening();
-
-        return result;
-    }
-
     public disconnect(): void {
         this.gateway.disconnect();
-    }
-
-    public async getMetadata(instantiatedChaincodeName: string, channel: string): Promise<any> {
-        const network: Network = await this.gateway.getNetwork(channel);
-        const smartContract: Contract = network.getContract(instantiatedChaincodeName);
-
-        let metadataBuffer: Buffer;
-        try {
-            metadataBuffer = await smartContract.evaluateTransaction('org.hyperledger.fabric:GetMetadata');
-        } catch (error) {
-            // This is the most likely case; smart contract does not support metadata.
-            throw new Error(`Transaction function "org.hyperledger.fabric:GetMetadata" returned an error: ${error.message}`);
-        }
-        const metadataString: string = metadataBuffer.toString();
-        if (!metadataString) {
-            // This is the unusual case; the function name is ignored, or accepted, but an empty string is returned.
-            throw new Error(`Transaction function "org.hyperledger.fabric:GetMetadata" did not return any metadata`);
-        }
-        try {
-            const metadataObject: any = JSON.parse(metadataBuffer.toString());
-
-            console.log('Metadata object is:', metadataObject);
-            return metadataObject;
-        } catch (error) {
-            // This is another unusual case; the function name is ignored, or accepted, but non-JSON data is returned.
-            throw new Error(`Transaction function "org.hyperledger.fabric:GetMetadata" did not return valid JSON metadata: ${error.message}`);
-        }
-    }
-
-    public async submitTransaction(chaincodeName: string, transactionName: string, channel: string, args: Array<string>, namespace: string, evaluate?: boolean): Promise<string | undefined> {
-        const network: Network = await this.gateway.getNetwork(channel);
-        const smartContract: Contract = network.getContract(chaincodeName, namespace);
-
-        let response: Buffer;
-        if (evaluate) {
-            response = await smartContract.evaluateTransaction(transactionName, ...args);
-        } else {
-            response = await smartContract.submitTransaction(transactionName, ...args);
-        }
-
-        if (response.buffer.byteLength === 0) {
-            // If the transaction returns no data
-            return undefined;
-        } else {
-            // Turn the response into a string
-            const result: any = response.toString('utf8');
-            return result;
-        }
-
-    }
-
-    public async upgradeChaincode(name: string, version: string, channelName: string, fcn: string, args: Array<string>): Promise<any> {
-
-        const transactionId: Client.TransactionId = this.gateway.getClient().newTransactionID();
-        const upgradeRequest: Client.ChaincodeInstantiateUpgradeRequest = {
-            chaincodeId: name,
-            chaincodeVersion: version,
-            txId: transactionId,
-            fcn: fcn,
-            args: args
-        };
-
-        const network: Network = await this.gateway.getNetwork(channelName);
-        const channel: Client.Channel = network.getChannel();
-
-        const instantiatedChaincode: Array<any> = await this.getInstantiatedChaincode(channelName);
-
-        const foundChaincode: any = this.getChaincode(name, instantiatedChaincode);
-
-        let proposalResponseObject: Client.ProposalResponseObject;
-
-        let message: string;
-
-        if (foundChaincode) {
-            message = `Upgrading with function: '${fcn}' and arguments: '${args}'`;
-            this.outputAdapter.log(LogType.INFO, undefined, message);
-            proposalResponseObject = await channel.sendUpgradeProposal(upgradeRequest);
-        } else {
-            //
-            throw new Error('The contract you tried to upgrade with has no previous versions instantiated');
-        }
-
-        const contract: Contract = network.getContract(name);
-        const transaction: any = (contract as any).createTransaction('dummy');
-
-        const responses: any = transaction['_validatePeerResponses'](proposalResponseObject[0]);
-
-        const txId: any = transactionId.getTransactionID();
-        const eventHandlerOptions: any = (contract as any).getEventHandlerOptions();
-        const eventHandler: any = transaction['_createTxEventHandler'](txId, network, eventHandlerOptions);
-
-        if (!eventHandler) {
-            throw new Error('Failed to create an event handler');
-        }
-
-        await eventHandler.startListening();
-
-        const transactionRequest: Client.TransactionRequest = {
-            proposalResponses: proposalResponseObject[0] as Client.ProposalResponse[],
-            proposal: proposalResponseObject[1],
-            txId: transactionId
-        };
-
-        // Submit the endorsed transaction to the primary orderers.
-        const response: Client.BroadcastResponse = await network.getChannel().sendTransaction(transactionRequest);
-
-        if (response.status !== 'SUCCESS') {
-            const msg: string = `Failed to send peer responses for transaction ${transactionId.getTransactionID()} to orderer. Response status: ${response.status}`;
-            eventHandler.cancelListening();
-            throw new Error(msg);
-        }
-
-        await eventHandler.waitForEvents();
-        // return the payload from the invoked chaincode
-        let result: any = null;
-        if (responses && responses.validResponses[0].response.payload.length > 0) {
-            result = responses.validResponses[0].response.payload;
-        }
-
-        eventHandler.cancelListening();
-
-        return result;
-    }
-
-    public async enroll(enrollmentID: string, enrollmentSecret: string): Promise<{certificate: string, privateKey: string}> {
-        const enrollment: ClientCA.IEnrollResponse = await this.gateway.getClient().getCertificateAuthority().enroll({ enrollmentID, enrollmentSecret });
-        return { certificate: enrollment.certificate, privateKey: enrollment.key.toBytes() };
-    }
-
-    public async getAllOrdererNames(): Promise<Array<string>> {
-
-        const ordererSet: Set<string> = new Set();
-        const allPeerNames: Array<string> = this.getAllPeerNames();
-
-        for (const peer of allPeerNames) {
-            const channels: string[] = await this.getAllChannelsForPeer(peer);
-            for (const _channelName of channels) {
-
-                const channel: Client.Channel = await this.getChannel(_channelName);
-                const orderers: Client.Orderer[] = channel.getOrderers();
-
-                for (const orderer of orderers) {
-                    ordererSet.add(orderer.getName());
-                }
-            }
-        }
-
-        return Array.from(ordererSet);
-    }
-
-    public async register(enrollmentID: string, affiliation: string): Promise<string> {
-        const request: ClientCA.IRegisterRequest = {
-            enrollmentID: enrollmentID,
-            affiliation: affiliation,
-            role: 'client'
-        };
-        const registrar: Client.User = this.gateway.getCurrentIdentity();
-        const secret: string = await this.gateway.getClient().getCertificateAuthority().register(request, registrar);
-        return secret;
     }
 
     public async createChannelMap(): Promise<Map<string, Array<string>>> {
@@ -425,37 +119,7 @@ export abstract class FabricConnection {
         }
     }
 
-    public async getAllInstantiatedChaincodes(): Promise<Array<{name: string, version: string}>> {
-
-        try {
-            const channelMap: Map<string, Array<string>> = await this.createChannelMap();
-            const channels: Array<string> = Array.from(channelMap.keys());
-
-            const chaincodes: Array<{name: string, version: string}> = []; // We can change the array type if we need more detailed chaincodes in future
-
-            for (const channel of channels) {
-                const channelChaincodes: Array<{name: string, version: string}> = await this.getInstantiatedChaincode(channel); // Returns channel chaincodes
-                for (const chaincode of channelChaincodes) { // For each channel chaincodes, push it to the 'chaincodes' array if it doesn't exist
-
-                    const alreadyExists: boolean = chaincodes.some((_chaincode: {name: string, version: string}) => {
-                        return _chaincode.name === chaincode.name && _chaincode.version === chaincode.version;
-                    });
-                    if (!alreadyExists) {
-                        chaincodes.push(chaincode);
-                    }
-                }
-            }
-
-            return chaincodes;
-        } catch (error) {
-            throw new Error(`Could not get all instantiated chaincodes: ${error}`);
-        }
-
-    }
-
     protected async connectInner(connectionProfile: object, wallet: FileSystemWallet, identityName: string): Promise<void> {
-
-        this.networkIdProperty = (connectionProfile['x-networkId'] ? true : false);
 
         this.discoveryAsLocalhost = this.hasLocalhostURLs(connectionProfile);
         this.discoveryEnabled = !this.discoveryAsLocalhost;
@@ -478,6 +142,36 @@ export abstract class FabricConnection {
 
         // TODO: remove this?
         this.mspid = identity.mspId;
+    }
+
+    protected getPeer(name: string): Client.Peer {
+        console.log('getPeer', name);
+        const allPeers: Array<Client.Peer> = this.getAllPeers();
+
+        return allPeers.find((peer: Client.Peer) => {
+            return peer.getName() === name;
+        });
+    }
+
+    protected async getChannel(channelName: string): Promise<Client.Channel> {
+        console.log('getChannel', channelName);
+        const client: Client = this.gateway.getClient();
+        let channel: Client.Channel = client.getChannel(channelName, false);
+        if (channel) {
+            return channel;
+        }
+        channel = client.newChannel(channelName);
+        const peers: Client.Peer[] = this.getAllPeers();
+        let lastError: Error = new Error(`Could not discover information for channel ${channelName} from known peers`);
+        for (const target of peers) {
+            try {
+                await channel.initialize({ asLocalhost: this.discoveryAsLocalhost, discover: this.discoveryEnabled, target });
+                return channel;
+            } catch (error) {
+                lastError = error;
+            }
+        }
+        throw lastError;
     }
 
     private isLocalhostURL(url: string): boolean {
@@ -506,42 +200,9 @@ export abstract class FabricConnection {
         return urls.some((url: string) => this.isLocalhostURL(url));
     }
 
-    private async getChannel(channelName: string): Promise<Client.Channel> {
-        console.log('getChannel', channelName);
-        const client: Client = this.gateway.getClient();
-        let channel: Client.Channel = client.getChannel(channelName, false);
-        if (channel) {
-            return channel;
-        }
-        channel = client.newChannel(channelName);
-        const peers: Client.Peer[] = this.getAllPeers();
-        let lastError: Error = new Error(`Could not discover information for channel ${channelName} from known peers`);
-        for (const target of peers) {
-            try {
-                await channel.initialize({ asLocalhost: this.discoveryAsLocalhost, discover: this.discoveryEnabled, target });
-                return channel;
-            } catch (error) {
-                lastError = error;
-            }
-        }
-        throw lastError;
-    }
-
     private getAllPeers(): Array<Client.Peer> {
         console.log('getAllPeers');
 
         return this.gateway.getClient().getPeersForOrg(this.mspid);
-    }
-
-    /**
-     * Get a chaincode from a list of list of chaincode
-     * @param name {String} The name of the chaincode to find
-     * @param chaincodeArray {Array<any>} An array of chaincode to search
-     * @returns {any} Returns a chaincode from the given array where the name matches the users input
-     */
-    private getChaincode(name: string, chaincodeArray: Array<any>): any {
-        return chaincodeArray.find((chaincode: any) => {
-            return chaincode.name === name;
-        });
     }
 }
